# client/exploitfarm/reqs/openapi-down.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FastAPI",
    "version": "{{VERSION_PLACEHOLDER}}"
  },
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login Api",
        "description": "Get a login token to use the apis",
        "operationId": "login_api_api_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_api_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse_Any_"
                }
              }
            }
          }
        }
      }
    },
    "/api/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get Status",
        "description": "This will return the application status, and the configuration if the user is allowed to see it",
        "operationId": "get_status_api_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusAPI"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse_Any_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/setup": {
      "post": {
        "tags": [
          "Status"
        ],
        "summary": "Set Status",
        "description": "Set some configuration values, you can set the values to change only",
        "operationId": "set_status_api_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse_Configuration_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse_Any_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttackMode": {
        "type": "string",
        "enum": [
          "wait-for-time-tick",
          "tick-delay",
          "loop-delay"
        ],
        "title": "AttackMode"
      },
      "Body_login_api_api_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "password"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_api_login_post"
      },
...
      "Configuration": {
        "properties": {
          "FLAG_REGEX": {
            "type": "string",
            "title": "Flag Regex",
            "default": ""
          },
          "START_TIME": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "END_TIME": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "TICK_DURATION": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "title": "Tick Duration",
            "default": 120
          },
          "ATTACK_MODE": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AttackMode"
              }
            ],
            "default": "tick-delay"
          },
          "LOOP_ATTACK_DELAY": {
            "type": "integer",
            "minimum": 0,
            "title": "Loop Attack Delay",
            "default": 0
          },
          "ATTACK_TIME_TICK_DELAY": {
            "type": "integer",
            "minimum": 0,
            "title": "Attack Time Tick Delay",
            "default": 0
          },
          "FLAG_TIMEOUT": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag Timeout"
          },
          "FLAG_SUBMIT_LIMIT": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Flag Submit Limit"
          },
          "SUBMIT_DELAY": {
            "type": "integer",
            "minimum": 0,
            "title": "Submit Delay",
            "default": 0
          },
          "SUBMITTER": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmitterID"
              },
              {
                "type": "null"
              }
            ]
          },
          "SUBMITTER_TIMEOUT": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitter Timeout",
            "default": 30
          },
          "AUTHENTICATION_REQUIRED": {
            "type": "boolean",
            "title": "Authentication Required",
            "default": false
          },
          "PASSWORD_HASH": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Hash"
          },
          "SETUP_STATUS": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SetupStatus"
              }
            ],
            "default": "setup"
          }
        },
        "type": "object",
        "title": "Configuration"
      },
...
      "MessageInfo": {
        "properties": {
          "level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageStatusLevel"
              }
            ],
            "default": "warning"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "title",
          "message"
        ],
        "title": "MessageInfo"
      },
...
      "MessageResponse_Any_": {
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseStatus"
              }
            ],
            "default": "ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "response": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Response"
          }
        },
        "type": "object",
        "title": "MessageResponse[Any]"
      },
...
      "MessageResponse_Configuration_": {
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseStatus"
              }
            ],
            "default": "ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Configuration"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MessageResponse[Configuration]"
      },
...
      "MessageStatusLevel": {
        "type": "string",
        "enum": [
          "info",
          "warning",
          "error"
        ],
        "title": "MessageStatusLevel"
      },
      "ResponseStatus": {
        "type": "string",
        "enum": [
          "ok",
          "error",
          "invalid"
        ],
        "title": "ResponseStatus"
      },
...
      "SetupStatus": {
        "type": "string",
        "enum": [
          "setup",
          "running"
        ],
        "title": "SetupStatus"
      },
      "StatusAPI": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/SetupStatus"
          },
          "loggined": {
            "type": "boolean",
            "title": "Loggined"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Configuration"
              },
              {
                "type": "null"
              }
            ]
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "auth_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Key"
          },
          "submitter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmitterDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "teams": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TeamDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teams"
          },
          "messages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MessageInfo"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "status",
          "loggined",
          "server_time",
          "messages"
        ],
        "title": "StatusAPI"
      },
...
      "SubmitterDTO": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SubmitterID"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "format": "binary",
            "title": "Code"
          },
          "kargs": {
            "additionalProperties": {
              "type": "object"
            },
            "type": "object",
            "title": "Kargs",
            "default": {}
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "code",
          "created_at"
        ],
        "title": "SubmitterDTO"
      },
...
      "SubmitterID": {
        "type": "integer"
      },
...
      "TeamDTO": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/TeamID"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "short_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Name"
          },
          "host": {
            "type": "string",
            "title": "Host"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "short_name",
          "host",
          "created_at"
        ],
        "title": "TeamDTO"
      },
...
      "TeamID": {
        "type": "integer"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}